refactor(bot): split pressure code to its scope #22
